+++ src/main/java/nl/elastique/poetry/annotations/package-info.java
@@ -1,4 +1,0 @@
-/**
- * This holds all core annotations (JSON-parsing annotations can be found in nl.elastique.poetry.json.annotations)
- */
-package nl.elastique.poetry.annotations;