+++ src/main/java/poetry/annotations/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This holds all core annotations (JSON-parsing annotations can be found in poetry.json.annotations)
+ */
+package poetry.annotations;