+++ src/main/java/poetry/json/JsonPathException.java
@@ -1,0 +1,14 @@
+package poetry.json;
+
+public class JsonPathException extends Exception
+{
+    public JsonPathException(String message)
+    {
+        super(message);
+    }
+
+    public JsonPathException(String message, Throwable parent)
+    {
+        super(message, parent);
+    }
+}